+++ sigenergy2mqtt/devices/plant/ess_preheating.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from datetime import timezone
+
 import sigenergy2mqtt.sensors.plant_ess_preheating_read_write as rw
 from sigenergy2mqtt.common import DeviceType, Protocol
 from sigenergy2mqtt.devices import ModbusDevice
@@ -12,28 +14,29 @@ class ESSPreHeating(ModbusDevice):
         device_type: DeviceType,
         protocol_version: Protocol,
     ):
-        name = "Sigenergy ESS Pre-Heating"
+        name = "Sigenergy Plant ESS Pre-Heating"
         plant_suffix = "" if plant_index == 0 else str(plant_index + 1)
         super().__init__(device_type, name, plant_index, 247, "ESS Pre-Heating", protocol_version, plant_suffix=plant_suffix)
 
     @classmethod
-    async def create(cls, plant_index: int, device_type: DeviceType, protocol_version: Protocol) -> "ESSPreHeating":
+    async def create(cls, plant_index: int, device_type: DeviceType, tz: timezone, rated_charging_power: float, rated_discharging_power: float, protocol_version: Protocol) -> "ESSPreHeating":
         ess_preheating = ESSPreHeating(plant_index, device_type, protocol_version)
-        await ess_preheating._register_sensors()
+        await ess_preheating._register_sensors(rated_charging_power, rated_discharging_power, tz)
         return ess_preheating
 
-    async def _register_sensors(self) -> None:
+    async def _register_sensors(self, rated_charging_power: float, rated_discharging_power: float, tz: timezone) -> None:
+        mode = rw.ESSPreHeatingMode(self.plant_index)
         self._add_sensor(rw.ESSPreHeatingEnable(self.plant_index))
-        self._add_sensor(rw.ESSPreHeatingMode(self.plant_index))
-        self._add_sensor(rw.ESSPreHeatingAdvanceEnable(self.plant_index))
+        self._add_sensor(mode)
+        self._add_sensor(rw.ESSPreHeatingAdvanceEnable(self.plant_index, mode))
 
         for tou_index in range(1, 31):
             start_address = 50003 + (tou_index - 1) * 6
             end_address = start_address + 2
             power_address = start_address + 4
 
-            self._add_sensor(rw.ESSPreHeatingTOUTimeStart(self.plant_index, tou_index, start_address))
-            self._add_sensor(rw.ESSPreHeatingTOUTimeEnd(self.plant_index, tou_index, end_address))
-            self._add_sensor(rw.ESSPreHeatingTOUTargetPower(self.plant_index, tou_index, power_address))
+            self._add_sensor(rw.ESSPreHeatingTOUTimeStart(self.plant_index, tou_index, start_address, tz))
+            self._add_sensor(rw.ESSPreHeatingTOUTimeEnd(self.plant_index, tou_index, end_address, tz))
+            self._add_sensor(rw.ESSPreHeatingTOUTargetPower(self.plant_index, tou_index, power_address, rated_charging_power, rated_discharging_power))
 
         self._add_sensor(rw.ESSPreHeatingReservedSOC(self.plant_index))